+++ src/endpoint/album.rs
@@ -3,12 +3,13 @@ use serde_json::json;
 
 use crate::{
     auth::AuthFlow,
+    client::Body,
     model::{
         album::{Album, Albums, PagedAlbums, SavedAlbum, SimplifiedAlbum},
         track::SimplifiedTrack,
         Page,
     },
-    query_list, Result,
+    query_list, Nil, Result,
 };
 
 use super::{Builder, Endpoint};
@@ -123,15 +124,15 @@ impl<F: AuthFlow> Builder<'_, F, SavedAlbumsEndpoint> {
             .await
     }
 
-    pub async fn save<T: Serialize>(self, ids: &[T]) -> Result<()> {
+    pub async fn save<T: Serialize>(self, ids: &[T]) -> Result<Nil> {
         self.spotify
-            .put("/me/albums".to_owned(), json!({ "ids": ids }))
+            .put("/me/albums".to_owned(), Body::Json(json!({ "ids": ids })))
             .await
     }
 
-    pub async fn remove<T: Serialize>(self, ids: &[T]) -> Result<()> {
+    pub async fn remove<T: Serialize>(self, ids: &[T]) -> Result<Nil> {
         self.spotify
-            .delete("/me/albums".to_owned(), json!({ "ids": ids }))
+            .delete("/me/albums".to_owned(), Body::Json(json!({ "ids": ids })))
             .await
     }