+++ src/model/track.rs
@@ -4,12 +4,10 @@ use serde::Deserialize;
 use super::{album::SimplifiedAlbum, artist::SimplifiedArtist, *};
 
 #[derive(Clone, Debug, Deserialize)]
-#[serde(deny_unknown_fields)]
 pub struct Track {
     pub album: SimplifiedAlbum,
     pub artists: Vec<SimplifiedArtist>,
-    #[serde(default)]
-    pub available_markets: Vec<String>,
+    pub available_markets: Option<Vec<String>>,
     pub disc_number: u32,
     pub duration_ms: u32,
     pub explicit: bool,
@@ -35,7 +33,6 @@ pub(crate) struct Tracks {
 }
 
 #[derive(Clone, Debug, Deserialize)]
-#[serde(deny_unknown_fields)]
 pub struct SimplifiedTrack {
     pub artists: Vec<SimplifiedArtist>,
     #[serde(default)]
@@ -58,14 +55,12 @@ pub struct SimplifiedTrack {
 }
 
 #[derive(Clone, Debug, Deserialize)]
-#[serde(deny_unknown_fields)]
 pub struct SavedTrack {
     pub added_at: DateTime<Utc>,
     pub track: Vec<Track>,
 }
 
 #[derive(Clone, Debug, Deserialize)]
-#[serde(deny_unknown_fields)]
 pub struct LinkedFrom {
     pub external_urls: ExternalUrls,
     pub href: String,